setup  logout for jobseeker
